requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.0
python-dotenv==0.14.0
regex==2020.10.23
requests==2.24.0
requests-cache==0.5.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
